added dynamic routing for tags and categories.
