+++ src/app/tag/[id]/not-found.tsx
@@ -1,0 +1,10 @@
+const NotFound = () => {
+  return (
+    <div>
+      <h1>404 - Page Not Found</h1>
+      <p>Sorry, the page you are looking for does not exist.</p>
+    </div>
+  );
+};
+
+export default NotFound;
